Use select to determine what executable definition to use
When running on my OSX laptop, for some reason it was always selecting
the Linux binary.  I was looking at trying to fix this by adjusting
the logic, but I have seen other patterns using `select` and a
condition around the target OS/Arch and this feels cleaner to me.

The upstream conditions being depended on are defined here.
https://github.com/bazelbuild/bazel/blob/1c29dff364fd34d7e6158c812cfd6d96f66be747/src/conditions/BUILD#L87-L94