chore: package the examples with the distribution tarball
This is a requirement by the BCR, but at the same time it allows
us to also run the integration test using a single self contained
package, which should only require bazel to be run.

This adds the integration testing dev dependency but it is only
used to populate bazelrc file to be able to get the example files.